Hack to rename folders - step 1
