+++ src/walletSdk/index.ts
@@ -3,10 +3,10 @@ import axios, { AxiosInstance, AxiosRequestConfig } from "axios";
 
 import { Anchor } from "./Anchor";
 import { WalletSigner, DefaultSigner } from "./Auth/WalletSigner";
-import { Stellar } from "./Horizon/Stellar";
-import { NETWORK_URLS } from "./Horizon/constants";
+import { Stellar } from "./Horizonn/Stellar";
+import { NETWORK_URLS } from "./Horizonn/constants";
 import { Recovery } from "./Recovery";
-import { getUrlDomain } from "./Util/url";
+import { getUrlDomain } from "./Utill/url";
 
 /* tslint:disable-next-line:no-var-requires */
 const version = require("../../package.json").version;
